Refactor Docker label handling and harden allowlist refreshes
